requirements.txt
pandas
numpy
tensorflow
keras
scikit-learn
matplotlib
seaborn
joblib
PyYAML
streamlit